+++ koharness.php
@@ -1,14 +1,14 @@
 <?php
 /**
  * Configuration for the koharness module testing environment.
- * 
+ *
  * @author    Andrew Coulton <andrew@ingenerator.com>
  * @copyright 2013 inGenerator Ltd
  * @link      https://github.com/ingenerator/koharness
  */
-return array(
-  'modules' => array(
-    'kohana-doctrine2' => __DIR__,
-    'unittest'    => __DIR__.'/modules/unittest',
-  )
-);
+return [
+    'modules' => [
+        'kohana-doctrine2' => __DIR__,
+    ],
+    'syspath' => __DIR__.'/vendor/ingenerator/kohana-core',
+];